backend/requirements.txt
# ── Core Django ───────────────────────────────────────────────
Django==4.2.7
asgiref==3.11.0
sqlparse==0.5.5

# ── Django REST Framework ──────────────────────────────────────
djangorestframework==3.14.0
django-cors-headers==4.3.0
dj-database-url==3.0.1
djangorestframework-simplejwt==5.5.1

# ── API Documentation (OpenAPI 3.0 / ReDoc) ───────────────────
drf-spectacular==0.29.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.36.2
rpds-py==0.27.1
attrs==25.4.0
inflection==0.5.1
uritemplate==4.2.0

# ── OpenAI GPT-4o-mini ────────────────────────────────────────
openai==2.16.0
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
jiter==0.12.0
distro==1.9.0

# ── OpenFoodFacts + HTTP ──────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
soupsieve==2.8.3

# ── Database ──────────────────────────────────────────────────
psycopg2-binary==2.9.11

# ── Production Server ─────────────────────────────────────────
gunicorn==23.0.0
whitenoise==6.11.0

# ── Image handling (Pillow used by Django internals/admin) ────
Pillow==10.1.0

# ── Config / Utilities ────────────────────────────────────────
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3

# ── Pydantic (required by openai SDK) ────────────────────────
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2
anyio==4.12.1
sniffio==1.3.1
exceptiongroup==1.3.1
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
tqdm==4.67.1
